yestoday_constructor_web/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-bootstrap5==22.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.4.4
django-zoom-meetings==0.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
zope.event==4.6
zope.interface==5.5.2
